Add files via upload
Sure, here is a possible GitHub description for your Simon Says game created using HTML, CSS, and JavaScript:

This is a simple Simon Says game created using HTML, CSS, and JavaScript. The game randomly generates a sequence of four colors, and the player must repeat the sequence by clicking on the corresponding buttons. The game gets more difficult as the sequence length increases.

The game is designed to be a fun and challenging way to test your memory and reflexes. It can be played by one or more players.

To play the game, simply click any key on keyboard. The game will then generate a sequence of four colors. Click on the corresponding buttons to repeat the sequence. If you get a sequence correct, you will advance to the next level. If you get a sequence wrong, the game will end. Then press any Key again and Restart the game. HAVE FUN!